Fix 'requires' is an invalid keyword argument for 'file.comment'
There looks to be a PR on GitHub to add the requires keyword to the
file.comment module but this isn't available yet.

Not the best practice, but will have to ensure the file.comment state is
listed after the installation of gitlab-runner (as this package will add
the gitlab-runner user and create the .bash_logout file to be commented
out).