Install Docker deb package using Ubuntu repository
Runners are currently failing to build some Alpine Liunx containers.
This appears to be caused by the Docker Salt forumla installing the
binary by downloading it directly from the internet, the binary
currently being downloaded is an outdated version.  The version
currently defined in the Docker Salt formula defaults.yaml is 19.03.9
which is a version from early 2019.

The solution is to install Docker via Docker's Ubuntu repositories which
means Docker will be updated when running apt upgrade.